RR-839 - Refactor overviewController into separate controllers
